+++ packages/performance/src/index.js
@@ -1,0 +1,15 @@
+import {getWebVitals} from './collect.js';
+
+export class performancePlugin {
+  constructor() {
+  }
+  core({reportData}){
+    getWebVitals((res)=>{
+      const data = {
+        type: 'performance',
+        ...res
+      };
+      reportData(data);
+    });
+  }
+}