+++ packages/core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @tmonitor/core
 
+## 1.0.3
+
+### Patch Changes
+
+- add some main features
+- Updated dependencies
+  - @web-tmonitor/performance@1.1.0
+  - @web-tmonitor/utils@1.0.2
+
 ## 1.0.2
 
 ### Patch Changes