+++ packages/performance/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @web-tmonitor/performance
 
+## 1.1.0
+
+### Minor Changes
+
+- add some main features
+
+### Patch Changes
+
+- Updated dependencies
+  - @web-tmonitor/utils@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes