+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @web-tmonitor/utils
 
+## 1.0.2
+
+### Patch Changes
+
+- add some main features
+
 ## 1.0.1
 
 ### Patch Changes